Better directory structure
Basically “solutions/[label]/[PR id, 5 digits] – [Name]”.

RT is barely over 6 digits, so more than 6 is definitely too much for
now. Python has a bit over 500 PEPs, so 4 digits could be
reasonable. However, the counter is incremented for every problem too,
not just for every solution (that's just how github works), so we
should double or triple the estimated amount. All things considered, 5
digits should be good enough for a long time, and as a bonus it makes
the numbers distinct from RT tickets, and also somewhat distinct from
Rakudo bugs because the id's will be padded with zeros.

The name is included because this way it is easier to find the
required document quickly (nobody is going to remember the numbers
anyway).

Not sure about spaces and “–”, but we can tweak that if there's an
issue.

The document will not define the directory structure and now it just
says that contributors should keep it consistent. I don't want to make
the document longer for such a tiny detail.

I decided to use solution id's instead of problem id's because one PR
can resolve more than one problem, and this way it's a bit more
straightforward.