Add CODEOWNERS file
Just testing the github feature. More about it here:

* https://docs.github.com/en/free-pro-team@latest/github/creating-cloning-and-archiving-repositories/about-code-owners

See this comment to understand why it may be useful:

* https://github.com/Raku/problem-solving/issues/241#issuecomment-719074196

Basically, it is supposed to require the review from whoever is
assigned to the label, which should make it easier to see when the PR
is actually ready for review by all of the reviewers. Not sure if
this is going to work, but let's see.

The CODEOWNERS file (just like the templates) needs to be kept in sync
with other files, and README.md is the source of truth.