Solution for Raku/doc/issues/1246
This PR proposes a solution to Raku/doc/issues/1246.  (I now realize
that issue was in the doc repo rather than the problem-solving one and
thus _technically_ I should open an issue in this repo and discuss the
matter here before opening a PR.  Since that issue has been active for
5 years and has generated the a similar discussion as it would have as
a problem-solving issue, I believe it can stand in for an issue in
this repo.  But if anyone objects, I can close this PR and do this
more by the book.)

Specifically, this PR would allow alternate documentation frontends
such as the [one prepared by
@Altai-man](https://github.com/Altai-man/docs.raku.org) to have a home
on the raku.org website (at `docs.$new-site-name.raku.org`).  It would
also provide a list of all such alternate frontends, some guidance
about creating new ones, and a note that an alternate frontend could
one day become the new primary doc site.  Please see the solution
document for full details.